refactor: isolate bleach code
